combined all backbone views, models, and collection code and deleted redundant files
