+++ app/public/javascripts/collections.js
@@ -7,4 +7,4 @@ var DayCollection = Backbone.Collection.extend({
 var BookingCollection = Backbone.Collection.extend({
   model: BookingModel,
   url: '/booking'
-})
+});
