+++ app/views/layout.hbs
@@ -78,9 +78,6 @@
     <!--backbone-->
     <script src="http://cdnjs.cloudflare.com/ajax/libs/backbone.js/0.9.2/backbone-min.js" type="text/javascript"></script>
     <script src="http://cdnjs.cloudflare.com/ajax/libs/backbone-localstorage.js/1.0/backbone.localStorage-min.js" type="text/javascript"></script>
-    <script type="text/javascript" src="../public/javascripts/views.js"></script>
-    <script type="text/javascript" src="../public/javascripts/models.js"></script>
-    <script type="text/javascript" src="../public/javascripts/main.js"></script>
     </head>
 
     <!-- Latest compiled and minified JavaScript -->
